improved bad channel (rfi) mask
